Alterations
i made some alterations to make to code look better for any people that is seeing the code, also added fetch helper that consist in only grab the data from the database. this helper are used on the viewModels of the project. the others helpers are used in different parts of the project. also added new app icon and contants for the image.